Rename customSecretKey to customSecretLicenseKey
The other New Relic charts like newrelic-infrastructure and
nri-prometheus use the name `customSecretLicenseKey`.

With these changes we are adding support for both options so that we
don't break users that are already using the old name. To encourage
usage we remove the old name and change it for the new one.