Changes to support running newrelic metadata injector in Kubernetes clusters with restrictive Pod Security Policy.
The changes to deployment.yaml allow you to configure a Pod Security Context for the main metadata injector pod so that it doesn't run as root and get denied scheduling
The changes to the job.yaml allow you to configure mounts in the Pod so that /tmp can be mounted as an emptydir when policy enforces a read-only root filesystem.
Additional Values added to values.yaml
README.md updated with new configuration values
Chart version bumped.