feat: add ns to sa
Should be more compatible to create the sa in the same ns as the minion.

Using the default sa in the default ns or even a synthetics-minion sa
in the default ns can sometimes lead to issues with sa permissions
depeding on how restricted the default ns is on the cluster.