feat: add resources to initContainer
If a resource quota has been defined for the namespace, the statefulset
will fail to create the initContainer unless we specify resource values for cpu and memory.

If the quota only defines hard limits, then only limits are needed on
the container spec. If the quota defines both hard request and limit
values, then both are needed on the initContainer spec. To cover all
cases we can specify both.