logging: fix empty messages with CRI
When running containerd, the log message was being dropped due 'message'
not being in the record_modifier allowlist. The cri parser now
unmarshals the field as 'log' instead of 'message'. 'log' is the
preferred field name as the Kubernetes filter checks that key for JSON
when Merge_Log is set to true.