[nri-bundle] manually update nri-bundle chart version (#1063)
#### Is this a new chart
No

#### What this PR does / why we need it:
A recent change to the nri-bundle Chart
(https://github.com/newrelic/helm-charts/pull/1061) did not trigger
dependency version updates from the renovate bot, so the nri-bundle
version stayed at 5.0.8. The release workflow seems to have recognized
this as a change and is trying to generate a new nri-bundle, but fails
because the bundle for 5.0.8 already exists. Subsequent PR merges fail,
presumably because the change is still detected. This change manually
updates the nri-bundle chart version in order to let the release
workflow package up the nri-bundle properly.

Slack discussion for internal NR for more context:
https://newrelic.slack.com/archives/C2XN6HL2G/p1681249969204639


#### Checklist
[Place an '[x]' (no spaces) in all applicable fields. Please remove
unrelated fields.]
- [x] Chart Version bumped
- [x] Title of the PR starts with chart name (e.g. `[mychartname]`)