fix: Update docs in response to GTSE.
ticket: https://new-relic.atlassian.net/jira/software/c/projects/NR/boards/500?selectedIssue=NR-170132

In response to the above GTSE, I am updating the README here to have the (Not yet configurable) note next to replica count. This is necessary to avoid confusion of a customer trying to change the replica count when we have not added support for that yet.