[newrelic-logging]: Remove unneeded hostNetwork setting (#1207)
#### Is this a new chart

No

#### What this PR does / why we need it:

We think that `hostNetwork: true` got copied-and-pasted from another
chart when the `newrelic-logging` was originally created a few years
ago. We want the `newrelic-logging` pod to have as little access as
possible, to increase security -- it should just need as much access as
it needs to do its job and no more. As far as we can tell, it doesn't
need to connect to the host network.

Also, PodSecurityPolicy is now deprecated and
[removed from the latest
versions](https://kubernetes.io/blog/2021/04/06/podsecuritypolicy-deprecation-past-present-and-future/)
of Kubernetes, but we're keeping it because we have customers are older
versions of Kubernetes. There is no direct replacement for a PSP, see
[here](https://kubernetes.io/blog/2021/04/06/podsecuritypolicy-deprecation-past-present-and-future/#what-does-this-mean-for-you)
for Kubernetes's advice.

#### Which issue this PR fixes

- Fixes #335 
  - Note: We do not think that we need to mount `/etc/hostname`

#### Special notes for your reviewer:

I've verified that the logs forwarded by the chart have the same
attributes (specifically, the same `host` value) before and after this
change. So this `hostNetwork` change _should_ be a no-op -- no
functionality will change (besides us not asking for the pod to have
access to the host network)

#### Checklist
[Place an '[x]' (no spaces) in all applicable fields. Please remove
unrelated fields.]
- [x] Chart Version bumped
- [x] Title of the PR starts with chart name (e.g. `[mychartname]`)