[nr-k8s-otel-collector] Honor the LowDataMode toogle from the common-library (#1472)
#### Is this a new chart
No

#### What this PR does / why we need it:
The value for `lowDataMode` was hard coded at values level so a user
cannot set the value if you not go deep down to the chart and patch it.

This change has its own implementation of the lowDataMode helper that
defaults to `true` while honoring the behaviour of the common-library of
setting a value a `global` level.

#### Which issue this PR fixes

#### Special notes for your reviewer:

#### Checklist
- [x] Chart Version bumped
- [x] Variables are documented in the README.md
- [x] Title of the PR starts with chart name (e.g. `[mychartname]`)

---------

Co-authored-by: csongnr <115833851+csongnr@users.noreply.github.com>
Co-authored-by: Daniel Budziwojski <dbudziwojski@newrelic.com>