[agent-control-deployment] improve security (#1785)
#### Is this a new chart

no

#### What this PR does / why we need it:

Improves the security of the container used in the preinstallation job.
Moves the folder where the private key is stored inside `/tmp`. That
will relax privileges constraints. We won't need to use `root` user
anymore to create the folder.

#### Which issue this PR fixes

  - relates to #NR-424891

#### Special notes for your reviewer:

There's an ongoing effort to split namespaces in AC. We are using branch
https://github.com/newrelic/helm-charts/tree/feat/splitNamespace from
this repo to test that everything is still working. I created a
separated branch on top of it that includes the changes of this PR. You
can find it
https://github.com/newrelic/helm-charts/tree/poc-temp-folder. This is
the one I used to test that everything is still working.

#### Checklist
[Place an '[x]' (no spaces) in all applicable fields. Please remove
unrelated fields.]
- [x] Chart Version bumped
- [x] Variables are documented in the README.md
- [x] Title of the PR starts with chart name (e.g. `[mychartname]`)