[nr-k8s-otel-collector, nri-bundle] Resource defaults  (#2012)
#### Is this a new chart
It is NOT a new chart.

#### What this PR does / why we need it:
Adds defaults for resource requirements to prevent otel from logging
WARN messages when installing these charts.

#### Which issue this PR fixes
*(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)`
format, will close that issue when PR gets merged)*
N/A

#### Special notes for your reviewer:
This PR is aimed at ensuring that customers can always set resource
values. It adds a passthrough for the init container to inherit the same
limits as the collector that is booted after it runs. I've also added
empty resource config blocks for the kube-state-metrics sub-chart as
in-line documentation to let user's know that this is one of the config
options available to be passed to the kube state metrics chart.

Not only is it just good practise to ensure our customers have easy
access to these settings, but It's especially important for GKE
Autopilot users who get warnings and default limits set on them if they
are not set manually. Which is what drove this effort, so I've also
added a callout in the readme for that.

#####No Resource values set:
<img width="1492" height="233" alt="Screenshot 2025-11-18 at 11 50
01 AM"
src="https://github.com/user-attachments/assets/c61a6969-e121-46eb-a57b-bf2fd530d7f8"
/>

#####With Resource values set: 
<img width="710" height="125" alt="Screenshot 2025-11-18 at 11 48 38 AM"
src="https://github.com/user-attachments/assets/12f18a9b-cb8b-4465-ad01-1210e32d4b41"
/>

#### Checklist
[Place an '[x]' (no spaces) in all applicable fields. Please remove
unrelated fields.]
- [x] Chart Version bumped
- [x] Variables are documented in the README.md
- [x] Title of the PR starts with chart name (e.g. `[mychartname]`)

# Release Notes to Publish (nr-k8s-otel-collector)
If this PR contains changes in `nr-k8s-otel-collector`, please complete
the following section. All other charts should ignore this section.

<!--BEGIN-RELEASE-NOTES-->
## 🚀 What's Changed
* Adding empty resource blocks for the Kube-state-metrics subchart in
the Values.yaml to appease providers that do not allow empty resource
blocks.
* Adds the user set resource values for the daemonset init container. 
<!--END-RELEASE-NOTES-->